Create GFG_125_Stock_Buy_and_Sell_Max_K_Transactions_Allowed.java
Approach: Use recursion and memoization to solve it. My approach is to use a boolean variable to check that the stack is already buy or not then we can decide that we have to sell or not. We should use memoization to reduce repeated recursive calls.

TC : O(n*k)
SC : O(k)