Update LeetCode_2033_Minimum_Operations_to_Make_a_Uni_Value_Grid.java
Approach: idea is to convert the 2d array to a 1d array and sort the array. Suppose two elements a and b belongs to the Array and a % x not equals to b % x then the array can never be converted into univalue grid. Then find the median of the array and count how many operation required to convert each element to reach that median value.

TC : O(n*m log n*m)
SC : O(n*m)