Create GFG_138_DFS_of_Graph.java
Approach: Use a variable to hold current vertex, a boolean array of size(no. of vertices) to check the node was visited or not and a arrayList to store the ans. Mark visited true to the current vertex and add it to ans list then call recursively for every neighbour of current vertex.

TC : O(V + E)
SC : O(V + E)