Update GFG_138_DFS_of_Graph.java
Approach: Use a variable to hold current vertex, a boolean array of size(no. of vertices) to check the node was visited or not and a arrayList to store the ans. Mark visited true to the current vertex and add it to ans list then call recursively for every neighbour of current vertex.

1. Main Method:
    ● Initializes the adjacency list adj.
    ● Calls the dfs method to perform DFS starting from vertex 0.
    ● Prints the result of the DFS traversal.

2. DFS Method:
    ● Initializes an ArrayList ans to store the DFS traversal result.
    ● Calls the private dfs helper method to start the DFS 
    ● traversal from vertex 0.

3. DFS Helper Method:
    ● Marks the current node as visited.
    ● Adds the current node to the result list.
    ● Recursively visits all unvisited neighbors of the current node.

TC : O(V + E)
SC : O(V + E)