Create GFG_139_BFS_of_graph.java
Approach: Use a queue to store neighbours level wise, ArrayList to store answer and a boolean array to check current vertex is visited or not. First create a queue and add a random vertex from which you want to start traversing. Run a loop until queue is not empty and in every iterations poll current vertex, mark that as visited and push all the neighbors of that vertex to the queue.

TC : O(V + E)
SC : O(V + E)