Update GFG_139_BFS_of_graph.java
Approach: Use a queue to store neighbours level wise, ArrayList to store answer and a boolean array to check current vertex is visited or not. First create a queue and add a random vertex from which you want to start traversing. Run a loop until queue is not empty and in every iterations poll current vertex, add current vertex to answer, push all the un-visited neighbors of that vertex to the queue and mark that neighbors as visited.

1. Import Statements: 
    The necessary libraries (ArrayList, LinkedList, List, and Queue) are imported.

2. Main Method:
    ● An adjacency list adj is created to represent the graph.
    ● The adjacency list is populated with the vertices and their respective connections.
    ● The bfs method is called with the adjacency list and the result is printed.

3. BFS Method (bfs):
    ● The size of the adjacency list is determined.
    ● An array visited is created to keep track of visited vertices.
    ● A queue q is created to manage the vertices to be processed.
    ● An ArrayList ans is used to store the BFS traversal result.
    ● The BFS traversal starts from vertex 0 and proceeds to visit all connected vertices.
    ● The algorithm adds vertices to the queue, marks them as visited, and adds them to the result list in the correct order.

TC : O(V + E)
SC : O(V + E)