Update GFG_144_Directed_Graph_Cycle.java
Approach: Use a boolean array to check if the visited node is a part of current path. BFS call for check if the cycle present, set every node true as back path and set false while backtrack. If the node is visited and also back path is true then that is a cycle.

—● Graph Representation: The graph is represented as an adjacency list.

—● Cycle Detection: The cycle detection is performed using Depth First Search (DFS). The visited array keeps track of visited nodes, and the backEdge array keeps track of nodes in the current path of DFS to detect cycles.

TC : O(v + e)
SC : O(v + e)