Create GFG_148_Dijkstra_Algorithm.java
Approach: Idea is to use a priority queue to store a pair of values containing vertex and weight and sort in ascending order in terms of weights. There is no visited array we have to traverse until all paths are found. Traversing is similar to BFS, run a loop until queue is not empty and poll a vertex visit all the neighbors of that vertex and update the distance for every neighbor which is possible and add the updated neighbor to the queue.

TC : O((v+e) log (v))
SC : O(v)