Create GFG_155_Implement_Trie.java
Approach: The Trie data structure is a tree-like structure used for storing a dynamic set of strings. It allows for efficient retrieval and storage of keys, making it highly effective in handling large datasets. Trie supports operations such as insertion, search, deletion of keys, and prefix searches. In this article, we will explore the insertion and search operations and prefix searches in Trie Data Structure.

TC :  ● Insertion : O(L)
        ● Search : O(L)
        ● Prefix : O(L)
SC :  ● Insertion : O(L)
        ● Search : O(1)
        ● Prefix : O(1)