Create Longest_Subarray_with_Majority_Greater_than_K.java
Approach: Idea is to use a HashMap to store the prefix sum occurs only first time and at which index. As the question says we have to calculate the majority element which strictly greater then k. We add 1 for greater element otherwise add -1 to the sum and store it in map with index. Check if sum-1 present in the map then calculate the distance. 
        It works because, think that sum is the prefix_sum from 0th_index to current_index, then if sum-1 present in the map then sum of one half elements is sum-1 so another half of elements sum must be sum+1, it means +1 occurs more the 0 times means that subarray must contain majority greater elements. 

TC : O(n)
SC : O(n)