Create Searching_Split_Array_Largest_Sum.java
Approach: Idea is to apply binary search to solve it. We need the minimized version of maximum subarray sum. So maximum sub-array sum might be sum of every elements if k is 1 or it might be largest element present in that array if k is equals with n. If I am right then the optimized subarray sum lies between in this. Then try binary search.

TC : O(n log total_sum)
SC : O(1)