Create S10_Sum_of_subarray_minimum.java
Approach: For each number calculate the number of sub-arrays which minimum is current number. So first we have to get Next smaller element and previous smaller element, then only we can calculate the next distance and previous distance and also can calculate the number of subarray with which current element is associated. Here is an edge case if an array containing multiple smaller elements like [2, 1, 3, 4, 1, 5] , here 1 occurs twice. In this case our approach could be include some sub-arrays twice or thrice. To overcome from this situation we should find the next smaller but previous smaller or equal element.

TC : O(n)
SC : O(n)