Create Sorting_Sort_by_Absolute_Difference.java
Approach: I know GFG mention to do it in constant space, but it's in java it's not possible because Comparator can't be used over primitive datatype arrays (e.g. int[] arr). So, we have to convert it into Integer Array before sorting.

TC : O(n)
SC : O(n)  ---> in java