Create Searching_Maximize_the_minimum_difference_between_k_elements.java
Approach: Idea is to use binary search on maximum possible answer and minimum possible answer. Then check if mid is a possible answer then try to maximize it. To check mid is a possible answer or not at linear time we have to sort the array.

TC : O(n log m)
SC : O(1)