Update Product_Modulo_Check.java
Approach: Idea is simple, we need to only handle right most digit, and a digit at most can be taken for 3 times bcz we have to pick 3 elements. So we take a array of size 10 and store the frequency of digits according to right most digit of each element. Then remake a list of digits containing respective values of elements and in worst case its size might be 30 and iteration might be 27000 so it takes constant time to traverse.

TC : O(n)
SC : O(1)