Create H_Sum_of_Mode.java
Approach: Here the main issue comes when we trying get the most frequent smaller element for running window. To achieve this functionality we have to use a combination of HashMap and PrioritieQueue. But the main issue is insertion and deletion operation on PriorityQueue takes linear time so we can use tree set to optimize it. tree set takes logarithmic time for this but we have to care fully deal with TreeSet sue to mutability issue.

TC : O(n log k)
SC : O(k)