Update S16_Max_of_min_for_every_window_size.java
Approach: try to find the current element's range means for which size of sub-array it is the minimum, update the max possible for that range. Un-filled ranges are must be filled with next filled range. One edge case, if for range 6 max of min is 50 then for range 5 or lesser then 6 can't be 40  lesser then 50, it should be 50 or greater then 50.

TC : O(n)
SC : O(n)