Create Recursion_54_All_Unique_Permutations_of_an_array.java
Approach: Idea is simple, we can get the next permutation of the array in O(n). So first we sort the array to get the smallest number can be formed by the array elements the call nextPermutation() until we get the largest number can be formed by using this array elements. OR we can use recursion also to get all the unique combinations.

TC : O(n! * n)
SC : O(n)