Create Graph_Minimum_Operations_to_Connect_Hospitals.java
Approach: Idea is simple, a graph only needs v-1 edges to connect all the vertices, so first we should return -1 if number of edges smaller than V-1. Otherwise we should count the number of connected components and return number of components - 1. It works because if there more edges than vertices in a graph then always there is a cycle or extra edge, So if there is a disconnected edge then it need a edge to connect.

TC : O(V + E)
SC : O(V + E)